[Part 1] Rename modules to use snake_case (#98)
* Rename reconciliation_visualization.py to recongraph_visualization.py

* Rename Diameter.py to diameter.py

* Rename DTLMedian.py to median.py

* Rename DTLReconGraph.py to recongraph_tools.py

* Rename ReconcileMainInput.py to reconcile_main_input.py

* Rename HistogramAlg.py to histogram_alg.py

* Rename HistogramAlg.py to histogram_alg.py

* Rename HistogramDisplay.py to histogram_display.py

* Rename HistogramMain.py to histogram_main.py

* Rename HistogramMainInput.py to histogram_main_input.py

* Delete unused input cli files

* Rename script02_gen_newick_trees.py to generate_newick_trees.py

* Remove unused tests

* Rename ClusterMain.py to cluster_main.py

* Remove unused input file

* Rename ClusterUtil.py to cluster_util.py

* Remove sampleDtlInputs folder (legacy from histogram codebase)

* Rename newickFormatReader.py to input_reader.py
