+++ empress/__init__.py
@@ -14,19 +14,19 @@ from typing import List
 from abc import ABC, abstractmethod
 
 from empress.xscape.CostVector import CostVector
-from empress import newickFormatReader
-from empress.newickFormatReader import ReconInput
-from empress.newickFormatReader import getInput as read_input
+from empress import input_reader
+from empress.input_reader import ReconInput
+from empress.input_reader import getInput as read_input
 from empress.xscape.reconcile import reconcile as xscape_reconcile
 from empress.xscape.plotcostsAnalytic import plot_costs_on_axis as xscape_plot_costs_on_axis
-from empress.reconcile import DTLReconGraph
-from empress.reconcile import reconciliation_visualization
-from empress.reconcile import DTLMedian
-from empress.reconcile import Diameter
+from empress.reconcile import recongraph_tools
+from empress.reconcile import recongraph_visualization
+from empress.reconcile import median
+from empress.reconcile import diameter
 from empress.reconcile import statistics
-from empress.histogram import HistogramDisplay
-from empress.histogram import HistogramAlg
-from empress.cluster import ClusterUtil
+from empress.histogram import histogram_display
+from empress.histogram import histogram_alg
+from empress.cluster import cluster_util
 from empress.recon_vis import recon_viewer
 
 def _find_roots(old_recon_graph) -> list:
@@ -115,19 +115,19 @@ class ReconGraphWrapper(Drawable):
         Draw Pairwise Distance Histogram on axes
         """
         # Reformat the host and parasite tree to use it with the histogram algorithm
-        gene_tree, gene_tree_root, gene_node_count = Diameter.reformat_tree(self.recon_input.parasite_tree, "pTop")
+        gene_tree, gene_tree_root, gene_node_count = diameter.reformat_tree(self.recon_input.parasite_tree, "pTop")
         species_tree, species_tree_root, species_node_count \
-            = Diameter.reformat_tree(self.recon_input.host_tree, "hTop")
-        hist = HistogramAlg.diameter_algorithm(
+            = diameter.reformat_tree(self.recon_input.host_tree, "hTop")
+        hist = histogram_alg.diameter_algorithm(
             species_tree, gene_tree, gene_tree_root, self.recongraph, self.recongraph,
             False, False)
-        HistogramDisplay.plot_histogram_to_ax(axes, hist.histogram_dict)
+        histogram_display.plot_histogram_to_ax(axes, hist.histogram_dict)
 
     def draw_graph_to_file(self, fname):
         """
         Draw self and save it as image at path fname.
         """
-        reconciliation_visualization.visualize_and_save(self.recongraph, fname)
+        recongraph_visualization.visualize_and_save(self.recongraph, fname)
 
     def stats(self, num_trials: int = 50):
         _, costs, p = statistics.stats(self.recon_input, self.dup_cost, self.trans_cost, self.loss_cost, num_trials)
@@ -147,16 +147,16 @@ class ReconGraphWrapper(Drawable):
         Return one of the best ReconciliationWrapper that best represents the
         reconciliation graph. The function internally uses random and is not deterministic.
         """
-        postorder_parasite_tree, gene_tree_root, _ = Diameter.reformat_tree(self.recon_input.parasite_tree, "pTop")
-        postorder_host_tree, _, _ = Diameter.reformat_tree(self.recon_input.host_tree, "hTop")
+        postorder_parasite_tree, gene_tree_root, _ = diameter.reformat_tree(self.recon_input.parasite_tree, "pTop")
+        postorder_host_tree, _, _ = diameter.reformat_tree(self.recon_input.host_tree, "hTop")
 
         # Compute the median reconciliation graph
-        median_reconciliation, n_meds, roots_for_median = DTLMedian.get_median_graph(
+        median_reconciliation, n_meds, roots_for_median = median.get_median_graph(
             self.recongraph, postorder_parasite_tree, postorder_host_tree, gene_tree_root, self.roots)
 
-        med_counts_dict = DTLMedian.get_med_counts(median_reconciliation, roots_for_median)
+        med_counts_dict = median.get_med_counts(median_reconciliation, roots_for_median)
 
-        random_median = DTLMedian.choose_random_median_wrapper(median_reconciliation, roots_for_median, med_counts_dict)
+        random_median = median.choose_random_median_wrapper(median_reconciliation, roots_for_median, med_counts_dict)
         median_root = _find_roots(random_median)[0]
         return ReconciliationWrapper(random_median, median_root, self.recon_input, self.dup_cost, self.trans_cost,
                                      self.loss_cost, self.total_cost)
@@ -169,15 +169,15 @@ class ReconGraphWrapper(Drawable):
             raise Exception("Cannot cluster %d Reconciliation into %d clusters" % (self.n_recon, n))
 
         gene_tree, species_tree, gene_root, recon_g, mpr_count, best_roots = \
-            ClusterUtil.get_tree_info(self.recon_input, self.dup_cost, self.trans_cost, self.loss_cost)
+            cluster_util.get_tree_info(self.recon_input, self.dup_cost, self.trans_cost, self.loss_cost)
 
-        score = ClusterUtil.mk_pdv_score(species_tree, gene_tree, gene_root)
+        score = cluster_util.mk_pdv_score(species_tree, gene_tree, gene_root)
 
-        graphs, scores, _ = ClusterUtil.cluster_graph(self.recongraph, gene_root, score, 4, n, 200)
+        graphs, scores, _ = cluster_util.cluster_graph(self.recongraph, gene_root, score, 4, n, 200)
         new_graphs = []
         for graph in graphs:
             roots = _find_roots(graph)
-            n = DTLReconGraph.count_mprs_wrapper(roots, graph)
+            n = recongraph_tools.count_mprs_wrapper(roots, graph)
             new_graphs.append(
                 ReconGraphWrapper(graph, roots, n, self.recon_input, self.dup_cost, self.trans_cost, self.loss_cost,
                                   self.total_cost))
@@ -217,5 +217,5 @@ def reconcile(recon_input: ReconInput, dup_cost: int, trans_cost: int, loss_cost
     Given recon_input (which has parasite tree, host tree, and tip mapping info)
     and the cost of the three events, computes and returns a reconciliation graph.
     """
-    graph, total_cost, n_recon, roots = DTLReconGraph.DP(recon_input, dup_cost, trans_cost, loss_cost)
+    graph, total_cost, n_recon, roots = recongraph_tools.DP(recon_input, dup_cost, trans_cost, loss_cost)
     return ReconGraphWrapper(graph, roots, n_recon, recon_input, dup_cost, trans_cost, loss_cost, total_cost)
