--- empress/reconcile/sampleDtlInputs/Readme.md
@@ -1,3 +1,0 @@
-Contains pickle files that when extracted have the format 
-(HostTree, ParasiteTree, LeafMapping, tlDUP_COST, TRANSFER_COST, LOSS_COST)
-Can be used with the DP function in DTLReconGraph.py
